+++ render/actions/user.actions.js
@@ -1,4 +1,5 @@
 import settingsActions from './settings.actions';
+import logActions from './log.actions';
 import request, { login, notify, logout } from './helper';
 
 export const USER_LOGIN = 'USER_LOGIN';
@@ -67,7 +68,9 @@ const fetchAvatar = id => (dispatch) => {
     const img = Buffer.from(data, 'binary')
       .toString('base64');
     dispatch({ type: USER_AVATAR, data: { size, img: `data:image;base64,${img}` } });
-  }).catch(() => {});
+  }).catch((e) => {
+    dispatch(logActions.add({ message: 'User has avatar ID but cannot fetch thumbnail', detail: e.message }));
+  });
 };
 
 export default {